removed initial model architecture, fully replaced with jan's
