+++ models/__init__.py
@@ -6,8 +6,5 @@ def get_model(name, **model_args):
     if name == 'cosmoflow':
         from .cosmoflow import build_model
         return build_model(**model_args)
-    elif name == 'cosmojan':
-        from .cosmojan import build_model
-        return build_model(**model_args)
     else:
         raise ValueError('Model %s unknown' % name)
